my bytes are off when I'm parsing, I'm going to explicitly return a np float 32 also going to trace the gps 3d fix

+++ sensors/processes/gpsCapture.py
@@ -60,7 +60,7 @@ def gps_capture(log_queue: queue.Queue, config: dict):
         if v is None:
             l.trace("to_float_or_nan: v is None: " + str(v))
             return np.nan
-        return float(v)
+        return np.float32(v)
 
 
     def ellipsoid_alt_km():
@@ -71,7 +71,7 @@ def gps_capture(log_queue: queue.Queue, config: dict):
         if geoid_alt is None:
             return np.nan
         ellipsoid_alt = msl_alt + geoid_alt
-        return ellipsoid_alt / 1000.0
+        return np.float32(ellipsoid_alt / 1000.0)
 
     def values_or_none(vals):
         if np.any(np.isnan(np.array(vals, dtype=float))):
